requirements.txt
pyTelegramBotAPI
MySQL-python
random2